BUG: Test added and error fixed in HDF5ImageIO
This commit solve the issue #3442 (https://issues.itk.org/jira/browse/ITK-3442)

First step: Try to identify the issue with adding a test in itkHDF5ImageIOTest
This lead us to a problem when we try to ReadImageInformation (read from the HDF5File) after WriteImageInformation (writing into the HDF5File).
This error is telling us that H5Dopen2() have failed and seems to be linked with VoxelData which doesn't exist.

To fix the error:
We must be sure that every files are closed before reading or writing.
And so, CloseDataSetFile and new DataSet have been added in ReadImageInformation and WriteImageInformation (before Read/Write).
Make sure that we close everything before doing something 'new'.

By the way, openDataSet is moved from Read inside ReadImageInformation and
createDataSet is moved from Write inside WriteImageInformation.

Change-Id: Ia38d49492587cc835bfada60c2b79ebb673f6b27